Deduplicate `tt_001` `ProcessEvent` function
